add PacketNameChangeTest and move random package
